+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # Automated Tourism
+## v0.03 05/04/2019 5:03:27 PM
+#### Changed
+- disable all buttons except priority
+
+#### Added
+- function ATsetButtonStatus(ref, state) to AT_2Panels
+- function ATtoggleTouristBoundary(rocket, state)
+
+#### Todo
+- clear stat variables
+- ip status section
+- status hint for amount of tourists waiting
+- on screen notifications
+- icon for status section
+
+--------------------------------------------------------
 ## v0.2 05/04/2019 3:45:23 AM
 
 #### Added
@@ -6,9 +22,12 @@
 - automated calculations for departures
 
 #### Todo
-- disable all ip buttons
-- clear stat variable
+- clear stat variables
 - ip status section
+- circle radius for landing
+- sign for rocket
+- status hint for amount of tourists waiting
+- on screen notifications
 
 --------------------------------------------------------
 ## v0.1 05/01/2019 1:45:02 AM